requirements.txt
fastapi
uvicorn
motor
python-dotenv
python-jose
bcrypt
pandas
python-multipart
pytest
httpx
